AI-Powered Resume Screener/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
pyodbc==4.0.32
python-dotenv==0.19.0
pdfplumber==0.6.2
docx2txt==0.8
textract==1.6.3
python-jose==3.3.0
werkzeug==2.0.1